requirements.txt
altair                  ==  5.5.0
attrs                     == 25.4.0
blinker                  ==  1.9.0
cachetools                 ==6.2.1
certifi             ==       2025.10.5
charset-normalizer     ==    3.4.4
click                     == 8.3.0
colorama         ==          0.4.6
gitdb               ==       4.0.12
GitPython              ==    3.1.45
idna                      == 3.11
Jinja2                ==     3.1.6
jsonschema               ==  4.25.1
jsonschema-specifications == 2025.9.1
MarkupSafe          ==       3.0.3
narwhals               ==    2.10.0
numpy                     == 2.3.4
packaging     ==             25.0
pandas           ==          2.3.3
pillow              ==        12.0.0
pip                     ==   25.0.1
protobuf                   ==6.33.0
pyarrow      ==              21.0.0
pydeck          ==           0.9.1
pygame             ==        2.6.1
python-dateutil       ==     2.9.0.post0
pytz                     ==  2025.2
referencing     ==           0.37.0
requests           ==        2.32.5
rpds-py               ==     0.28.0
six                      ==  1.17.0
smmap               ==       5.0.2
streamlit              ==    1.51.0
streamlit-autorefresh     == 1.0.1
tenacity                ==   9.1.2
toml                       ==0.10.2
tornado          ==          6.5.2
typing_extensions   ==       4.15.0
tzdata                 ==    2025.2
urllib3                   == 2.5.0
watchdog           ==        6.0.0
altair                  ==       5.5.0
annotated-types        ==        0.7.0
anyio                   ==       4.12.0
attrs                   ==       25.4.0
blinker                 ==       1.9.0
cachetools             ==        6.2.1
certifi                  ==      2025.10.5
charset-normalizer     ==        3.4.4
click              ==            8.3.0
colorama         ==              0.4.6
gitdb               ==           4.0.12
GitPython           ==           3.1.45
google-ai-generativelanguage ==  0.6.15
google-api-core       ==         2.28.1
google-api-python-client   ==    2.187.0
google-auth         ==           2.43.0
google-auth-httplib2   ==       0.2.1
google-genai       ==            1.52.0
google-generativeai    ==        0.8.5
googleapis-common-protos    ==   1.72.0
grpcio                  ==       1.76.0
grpcio-status               ==   1.71.2
h11                 ==           0.16.0
httpcore           ==            1.0.9
httplib2          ==            0.31.0
httpx             ==             0.28.1
idna                ==           3.11
Jinja2           ==              3.1.6
jsonschema       ==              4.25.1
jsonschema-specifications  ==    2025.9.1
MarkupSafe           ==          3.0.3
narwhals         ==              2.10.0
numpy           ==               2.3.4
packaging       ==               25.0
pandas            ==             2.3.3
pillow           ==              12.0.0
pip              ==              25.0.1
proto-plus      ==               1.26.1
protobuf         ==              5.29.5
pyarrow         ==               21.0.0
pyasn1          ==               0.6.1
pyasn1_modules     ==            0.4.2
pydantic          ==             2.12.5
pydantic_core     ==            2.41.5
pydeck          ==              0.9.1
pygame            ==             2.6.1
pyparsing       ==               3.2.5
python-dateutil      ==          2.9.0.post0
pytz         ==                  2025.2
referencing      ==              0.37.0
requests       ==               2.32.5
rpds-py        ==                0.28.0
rsa        ==                    4.9.1
six           ==                 1.17.0
smmap                 ==         5.0.2
streamlit            ==          1.51.0
streamlit-autorefresh  ==        1.0.1
tenacity           ==            9.1.2
toml             ==              0.10.2
tornado         ==              6.5.2
tqdm          ==                 4.67.1
typing_extensions     ==         4.15.0
typing-inspection    ==          0.4.2
tzdata        ==                 2025.2
uritemplate      ==              4.2.0
urllib3          ==              2.5.0
watchdog          ==             6.0.0
websockets             ==        15.0.1 